refactor: product, menu Aggregate 로 통합
