+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -11,8 +11,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
-import kitchenpos.product.domain.Product;
-
 @Entity
 public class MenuProduct {
