+++ src/main/java/kitchenpos/menu/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.menu.domain;
 
 import java.util.Objects;
 
@@ -8,8 +8,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-
-import kitchenpos.menu.domain.Price;
 
 @Entity
 public class Product {
