+++ src/main/java/kitchenpos/menu/repository/ProductRepository.java
@@ -1,9 +1,9 @@
-package kitchenpos.product.repository;
+package kitchenpos.menu.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
